update patching recommendations in FAQ
Looking a zmughal's use of the patching with perl pattern,
I realized the example in the FAQ could be much simpler if
we had the patch sub injected into main (do not have to
quote anything, or use the -> operator).  Also there really
isn't any reason that it needs to be in the AB::MB subclass
so I am just using My::AlienPatch instead without any subclassing
at all.