use travis new docker infrastructure
I noticed a huge speedup when I switched over to travis' new
container-based infrastructure, the main downside of which appears
to be that you cannot use sudo.  This required some alterations to
the way the tests were run, and I took the opportunity to break out
some of the work into separate .sh files.  Another upside here is
that we do not have to install dontpanic (the c lib) twice on
ALIEN_FORCE=1, as it was doing before.