build_uri: fix URI composition
Rely as much as possible into the URI module when composing two URIs.

Before this patch, the code used was too simplistic and only covered
the more basic cases. Specifically it failed to handle the links from
GitHub release pages.

Also, one of the tests in http_uri.t was wrong.