no warnings for dist with several .pc files
For packages providing multiple libraries and so, multiple .pc files,
it is possible for some keys to not be defined in all of them.
Specifically, OpenSSL doesn't defined "Libs" in all the .pc files.

This change just silences a warning that was generated at runtime when
looking for a keyword not available in some .pc file.