Test success on version probe
If the command fails, but there is some output we get a false positive.
This problem was discovered when using perl Build.PL verbose=1
(and demonstrated with Alien::LibYAML).  While this exact problem
will hopefully be fixed with PR #147, there are other strangeisms
that could infect the captured output with bogus data.