Use the base_url from the request
Briefly, the problem is that if you provide something like

 "/foo/bar"

as the location, and that location is a directory, then the remote
server will redirect you to "/foo/bar/" before returning the
directory listing.  LWP::UserAgent or HTTP::Tiny will do this for
you, but if you use "/foo/bar" as the base_url then any relative
URLs will be wrong.

Ideally AB modules should specify "/foo/bar/" but this is an easy
fix for modules that do not.