Use Alien::CMake 0.07 for Alien::Base compat
I added bin_dir method to Alien::CMake so it can be used like a regular
Alien::Base bases module with alien_bin_requires.  This simplifies the
environment code ever so slightly.  To improve reliability, I've also
added a minimum 0.07 requirement for Alien::CMake when an older version
or no version is specified.