Workaround for weird OpenBSD pkg-config
OpenBSD has its own implementation of pkg-config (implemented in
Perl as it happens).  It has many idiosyncrasies.  One is that
if you have a -lfoo flag without a -L/lib flag in Libs it will
put a leading space on it.  I am 99% sure this is harmless and
stripping the space will make the test pass on OpenBSD.